requirements.txt
torch
torchvision
numpy 
jupyter
ipykernel
matplotlib
natsort
pillow
tqdm